% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_filters.R
\name{define_filters}
\alias{define_filters}
\title{Helper function for the server-side of the app. Defines filters for the app.}
\usage{
define_filters(data_source)
}
\arguments{
\item{data_source}{information about data source.}
}
\description{
Helper function for the server-side of the app. Defines filters for the app.
}
