% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_custom_theme.R
\name{get_custom_theme}
\alias{get_custom_theme}
\title{Get custom CSS theme for shinydashboard/shinydashboardPlus projects}
\usage{
get_custom_theme(
  type = "bs4Dash-light-dark",
  font_family = "Source Sans Pro",
  ...
)
}
\arguments{
\item{type}{A type of theme to get. Available options: "shinydashboard-blue", "shinydashboard-blue-light", "bs4Dash-light-dark"}

\item{font_family}{Font family. Available options: "RocheSans", "RocheSansCondensed", "RocheSerif", "Source Sans Pro"}

\item{...}{Arguments to be passed through to [fresh::create_theme()].
These can be used to set the width, height, color, etc.}
}
\value{
Code of custom CSS theme
}
\description{
Get custom CSS theme for shinydashboard/shinydashboardPlus projects
}
\examples{
RocheIdentity::get_roche_theme()
RocheIdentity::get_roche_theme(type = "bs4Dash-light-dark", font_family = "Source Sans Pro")
}
