% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patients}
\alias{patients}
\title{Generated sample of Real World patients data}
\format{
An object of class \code{list} of length 4.
}
\usage{
patients
}
\description{
Generated sample of Real World patients data
}
\keyword{datasets}
