% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdemo2023-package.R
\docType{package}
\name{scdemo2023-package}
\alias{scdemo2023}
\alias{scdemo2023-package}
\title{scdemo2023: RWD IE Team Demo for ShinyConf 2023}
\description{
RWD IE Team Demo for ShinyConf 2023
}
\author{
\strong{Maintainer}: Krystian Igras \email{krystian8207@gmail.com}

}
\keyword{internal}
